Rename test so that surefire report of different modules do not collide
