+++ jzy3d-native-jogl-awt/src/test/java/org/jzy3d/painters/TestNativeDesktopPainter_AWT.java
@@ -3,16 +3,16 @@ package org.jzy3d.painters;
 import static org.mockito.Mockito.mock;
 import org.junit.Assert;
 import org.junit.Test;
-import org.jzy3d.chart.factories.CanvasNewtSWT;
 import org.jzy3d.os.WindowingToolkit;
+import org.jzy3d.plot3d.rendering.canvas.CanvasAWT;
 
-public class TestNativeDesktopPainter {
+public class TestNativeDesktopPainter_AWT {
   @Test
-  public void detectSwt() {
+  public void detectAWT() {
     NativeDesktopPainter p = new NativeDesktopPainter();
     
-    p.setCanvas(mock(CanvasNewtSWT.class));
+    p.setCanvas(mock(CanvasAWT.class));
     
-    Assert.assertEquals(WindowingToolkit.SWT, p.getWindowingToolkit());
+    Assert.assertEquals(WindowingToolkit.AWT, p.getWindowingToolkit());
   }
 }
