+++ jzy3d-native-jogl-swt/src/test/java/org/jzy3d/painters/TestNativeDesktopPainter_SWT.java
@@ -3,16 +3,16 @@ package org.jzy3d.painters;
 import static org.mockito.Mockito.mock;
 import org.junit.Assert;
 import org.junit.Test;
+import org.jzy3d.chart.factories.CanvasNewtSWT;
 import org.jzy3d.os.WindowingToolkit;
-import org.jzy3d.plot3d.rendering.canvas.CanvasAWT;
 
-public class TestNativeDesktopPainter {
+public class TestNativeDesktopPainter_SWT {
   @Test
-  public void detectAWT() {
+  public void detectSwt() {
     NativeDesktopPainter p = new NativeDesktopPainter();
     
-    p.setCanvas(mock(CanvasAWT.class));
+    p.setCanvas(mock(CanvasNewtSWT.class));
     
-    Assert.assertEquals(WindowingToolkit.AWT, p.getWindowingToolkit());
+    Assert.assertEquals(WindowingToolkit.SWT, p.getWindowingToolkit());
   }
 }
